Feat/AI calculated difficulty

+++ src/main/java/com/hangeulbada/domain/workbookset/dto/WorkbookDto.java
@@ -20,7 +20,7 @@ public class WorkbookDto {
     private String teacherId;
     private String title;
     private String description;
-    private double difficulty;
+    private Integer difficulty;
     private int questionNum;
     @Builder.Default
     private List<String> questionIds = new ArrayList<>();    private LocalDateTime startDate;
