+++ src/main/java/com/hangeulbada/domain/workbookset/entity/Question.java
@@ -17,7 +17,7 @@ public class Question {
     private String id;
     private String teacherId;
     private String content;
-    private double difficulty;
+    private Integer difficulty;
     private String[] tags;
     private String audioFilePath;
